Add spritesheet
`sprite.clj` contains an atlas that defines the coordinates and initial
state of the sprites found in `spritesheet.png`. It also contains a
macro which checks sprite names at compile time and throws an error if
they are not valid sprites.

`build/package_resources` converts and packages clojure resources into
the docs folder (could potentially move index.html in here).